requirements.txt
python-telegram-bot==20.7
gspread==5.12.4
oauth2client==4.1.3
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
python-dotenv==1.0.0
Flask==3.0.0
gunicorn==21.2.0
pytest==7.4.3
requests==2.31.0 